### 2017-10-24 #436 @whedon
[REVIEW]: Masks2Metrics (M2M): A Matlab Toolbox for Gold Standard Morphometrics
**Submitting author:** @smikhael (<a href="http://orcid.org/0000-0002-2490-9548">Shadia Mikhael</a>)
**Repository:** <a href="https://github.com/Edinburgh-Imaging/Masks2Metrics" target ="_blank">https://github.com/Edinburgh-Imaging/Masks2Metrics</a>
**Version:** v1.0
**Archive:** <a href="http://dx.doi.org/10.7488/ds/2302" target="_blank">10.7488/ds/2302</a>
