### 2017-10-24 Edinburgh-Imaging/Masks2Metrics#1 @Kevin-Mattheus-Moerman
Recommendation: Adhere to open source guidelines with respect to completeness of the readme, contributing.md, code of conduct, and roadmap
It would be good open up your project to other contributors more by having clear guidelines for contributing in a CONTRIBUTING.md file. In addition you can add a COC.md (code of conduct) file, and potentially a ROADMAP.md file to describe where the project is going.

See also here: https://opensource.guide/

For a code of conduct you can base it on this one:
https://www.contributor-covenant.org/version/1/4/code-of-conduct.html

For an example use of a contributing.md, coc.md, readme.md see here:
https://github.com/gibbonCode/GIBBON

The mozilla science group offers advice on how to set up projects. Here is some of their training material:
https://github.com/mozillascience/leadership-training
