### 2017-10-24 Edinburgh-Imaging/Masks2Metrics#2 @Kevin-Mattheus-Moerman
Add demos/documentation/examples and expand README.md to include "Getting started" with links to useful demos
Please expand the documentation/demos/examples such that new users have something to start from. There appears to be a zipped example folder (containing a example_copy folder, is that correct?) but new users do not know how to use this example.

Perhaps create a demo (or set of demos) for the toolbox such that others can see how to interact with and use the various aspects of the code more easily.

Next you create a "Getting started" section in your README.md file with links to these demos to guide the way. Here is an example of such a section in one of my toolboxes: https://github.com/gibbonCode/GIBBON#Start

Also in https://github.com/gibbonCode/GIBBON I have DEMO and HELP files which are "publishable" in MATLAB and this help and documentation can be integrated (and is searchable in MATLAB). If you also want this let me know and I can assist.
