### 2017-10-24 #436 @Kevin-Mattheus-Moerman
@smikhael I've started the review process. I've a couple of points before I dive any deeper.

I opened an issue in relation to [contributing guidelines](https://github.com/Edinburgh-Imaging/Masks2Metrics/issues/1) and one about [help/documentation and adding a "getting started" section to the readme.md](https://github.com/Edinburgh-Imaging/Masks2Metrics/issues/2).

The former issue is a recommendation, however the latter I feel is really required since (as a new user) I'm unsure where to start.
