### 2017-10-24 #436 @Kevin-Mattheus-Moerman
Is there a DOI for the Bauer_2013 reference? Perhaps remove the self citation in the paper (Mikhael_2017) to the github link of this project, or replace it by reference/DOI to an archive (e.g. ZENODO) once an initial release is created.
