### 2017-10-24 Edinburgh-Imaging/Masks2Metrics#2 @Kevin-Mattheus-Moerman
Woops, apologies I missed the wiki initially. So that partially covers this. I'll check it out shortly. Perhaps still add a getting started section to the readme though.
