### 2017-10-25 4de1835
Added doi and url for Bauer_2013 reference
