+++ paper for JOSS/paper.bib
@@ -6,7 +6,9 @@ month = {06},
 pages = {R97-R129},
 title = {A survey of MRI-based medical image analysis for brain tumor studies},
 volume = {58},
-booktitle = {Physics in medicine and biology}
+booktitle = {Physics in medicine and biology},
+doi = {10.1088/0031-9155/58/13/R97},
+url = {http://iopscience.iop.org/article/10.1088/0031-9155/58/13/R97/meta}
 }
 
 @article{Desikan_2006,
@@ -32,7 +34,7 @@ author = "Rahul S. Desikan and Florent Ségonne and Bruce Fischl and Brian T. Qu
   month = {APR},
   year = {2011},
   doi = {10.1212/wnl.0b013e3182166e96},
-  url = "https://www.ncbi.nlm.nih.gov/pubmed/21490323?dopt=Abstract"
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/21490323?dopt=Abstract}
 }
 
 @article{Fischl_2000,
