### 2017-10-25 13e8227
Added 'getting started' section
- included links to the wiki, tutorial, and workflow
