+++ README.md
@@ -10,13 +10,17 @@ This program is free software: you can redistribute it and/or modify it under th
 
 Masks2Metrics (M2M) is a Matlab based tool that is used to calculate 3 metrics for a given region-of-interest (ROI) in a 3D image: thickness, volume and suface area. While many software packages that automatically compute such metrics exist, it is also necessary to compare the results from automated software with manually-traced ROIs. The current software takes such manually-traced ROIs and computes the metrics automatically.
 
-We (the authors and contributors) use M2M to compute those metrics on images acquired by a Magnetic Resonance Imaging (MRI) machine. The ROI are defined by paired of masks (in NIfTI format) defining the inner and outer borders of the ROI, and they are drawn continuously along one direction (x-, y- or z-axis). For the special case of brain images, if the ROI describes a gyrus, the paired masks would be the corresponding grey matter (GM) and white matter curves (WM). Paired ROI NIfTI (.nii) masks are expected to be of the form subj_roi_hem_gm/wmsegments.nii. An example of a pair corresponding to subject 1 right superior frontal gyrus (SFG) would be 1_sfg_r_gm1.nii and 1_sfg_r_wm1.nii. A special feature of M2M is that multiple segments can be used rather than a single continuous ROI (see Wiki help). The ROI metrics outputted are grey matter thickness (GMth), grey matter volume (GMvol),and white matter surface area (WMsa).
+We (the authors and contributors) use M2M to compute those metrics on images acquired by a Magnetic Resonance Imaging (MRI) machine. The ROI are defined by paired of masks (in NIfTI format) defining the inner and outer borders of the ROI, and they are drawn continuously along one direction (x-, y- or z-axis). For the special case of brain images, if the ROI describes a gyrus, the paired masks would be the corresponding grey matter (GM) and white matter curves (WM). Paired ROI NIfTI (.nii) masks are expected to be of the form subj_roi_hem_gm/wmsegments.nii. An example of a pair corresponding to subject 1 right superior frontal gyrus (SFG) would be 1_sfg_r_gm1.nii and 1_sfg_r_wm1.nii. A special feature of M2M is that multiple segments can be used rather than a single continuous ROI (see Wiki help). The ROI metrics calculated are grey matter thickness (GMth), grey matter volume (GMvol),and white matter surface area (WMsa).
 
-## Installation instructions
+## Installation Instructions
 
 Download M2M into a folder called Masks2Metrics. In addition to the Masks2Metrics code, the folder includes external code that is called by the tool, including a NIfTI Matlab tools and pre-existing distance code (Euclidean, Fréchet and the Modified Hausdorff Distance).
 
-Add the Masks2Metrics folder and subfolders to your list of Matlab paths. To generate the metrics call masks2metrics as detailed in the Wiki.
+## Getting Started
+1. Add the Masks2Metrics folder and subfolders to your list of Matlab paths. 
+2. To calculate your ROI's metrics call the function <code>'masks2metrics'</code> with the appropriate parameters. 
+
+For more details, please refer to our _[Wiki](https://github.com/Edinburgh-Imaging/Masks2Metrics/wiki)_. It includes a _[short tutorial](https://github.com/Edinburgh-Imaging/Masks2Metrics/wiki/Short-tutorial)_ demonstrating how the tool can be used as well as a _[workflow](https://github.com/Edinburgh-Imaging/Masks2Metrics/wiki/Workflow)_ highlighting communication between all functions.
 
 ## Authors
 Shadia Mikhael, Edinburgh Imaging, Centre for Clinical Brain Sciences, University of Edinburgh, Edinburgh, UK.  
