### 2017-10-25 Edinburgh-Imaging/Masks2Metrics#2 @smikhael
No problem! I've just added a getting started section to the readme with links to the wiki (which isn't always obvious- it has happened to me many times). For the same reason, I've also added links to the short tutorial and workflow on the wiki homepage. Hope this helps.
