### 2017-10-25 #436 @smikhael
That's excellent. Thank you @cMadan  and @Kevin-Mattheus-Moerman for agreeing to be the editor and reviewer.
So far I have (1) added a doi and url for the Bauer_2013 reference (and will replace the self citation by a doi once an initial release is created, as suggested), and (2) added a 'getting started' section to the home page as well as links to the tutorial and workflow on both that page and the wiki home page, just so the user doesn't miss them. It is definitely off-putting when they don't see/have them.
I will next look into the contributing guidelines you recommended.
Thank you for looking into this project so promptly.
