### 2017-10-25 9a0db7e
Adding guidelines for contributors
Adding 'code of conduct' and 'contributing' files
