### 2017-10-25 eabe29c
Updated with TOC, license, and contributor info
- added table of contents
- added contributing and code of conduct sections and links to those files
- created separate license section and deleted accidentally repeated paragraph in that section
