### 2017-10-25 Edinburgh-Imaging/Masks2Metrics#1 @smikhael
I've now added the contributing and coc files, with reference to them in the README. The road map is 'coming soon'. A table of contents has been added to the READMEas well. Thank you for these recommendations.
