### 2017-10-25 Edinburgh-Imaging/Masks2Metrics#1 @Kevin-Mattheus-Moerman
:clap: Thanks, looks great.
