### 2017-10-25 Edinburgh-Imaging/Masks2Metrics#2 @Kevin-Mattheus-Moerman
:tada: thanks!
