### 2017-11-01 #436 @Kevin-Mattheus-Moerman
Hi @smikhael I'm still working on the review I should be able to finish it this week. If there are more issues I'll create them on your repo and mention them here.
