### 2017-12-13 #436 @CPernet
@Kevin-Mattheus-Moerman @cMadan we are planning for January to release some evaluations/comparisons of metrics against other software - is there anything else we should do for the repo/article? thx
