### 2017-12-22 #436 @Kevin-Mattheus-Moerman
@cMadan @CPernet @smikhael
I've completed the review and could verify that your software does _what it says on the can_. I have some minor comments below.

# About the wiki
The Wiki home link starts with
> Masks2Metrics (M2M) is a quick and robust Matlab-based tool that calculates 3 metrics for a given region-of-interest (ROI)- its thickness, its volume, and the surface area of the inner mask.

There may be different definitions of what an ROI is, e.g. ROI may be masks or are sometimes defined as closed polygon sets. I think in your case they are binary pixels sets that are continuous lines for each slice in a particular direction? Perhaps define what you mean by ROI. Could you also provide a link to the NIfTI file format definition or a link to tools that are able to open/create them?

I think you should probably remove the word `function` in the wiki (under home and usage), i.e. you would not run it with function like that as a command you probably accidentally pasted it in:
> After browsing to the directory containing the paired masks, the code is run using a single command line of the form:
  `function [roi_gm_mean_thickness, roi_gm_vol, roi_wm_sa] = masks2metrics(subj, segments, roi, hem, step_size, draw)`

# About the example
Not sure if this is serious / an issue only on my machine, but upon running the tutorial in the wiki I get this warning (I use MATLAB 2016b, Ubuntu 17.10):

> Warning: Could not start Excel server for export.
    XLSWRITE will attempt to write file in CSV format.
     > In xlswrite (line 174)
     In masks2metrics (line 209)

The example does not say where the files are saved. I found the file 1_sfg_l_mean_frechet_step3.mat in the current working directory. Could the user set the location? Why are these files saved separately like that? I also see csv files. Could you add a description of all data that is saved, where it is saved and what they represent, in the wiki? What can the user do with these files?

The following is not a requirement but a recommendation. The example rapidly generates ~27 figures. The images in these figures are very small and difficult to interpret. Also as a first example having 27 tiny figures splash across the screen seems a bit overwhelming/difficult to understand. The figures in the actual wiki (like figure 1 under home and figure 1 and 2 in the tutorial) are much clearer and contain a description. I.e. they are zoomed in on a feature and have a legend/description. Could you make the figures generated by the example very similar or actually have the example replicate these figures? Would you be able to create an anatomical view in matlab as well so the users can see what these masks represent? Also would it be possible to add legends to the figures to denote what the dots/symbols/lines mean? You could also explain in the wiki what the figures are going to show. So warn them about 27 figures E.g. by running this MATLAB will generate 27 figures representing each slice (?) and visualizing each mask and the distances metrics and areas between them?

# About the readme
There is a minor typo / an awkward sentence in the readme you should rephrase, (I think a similar statement is well written in the paper so you could perhaps copy it over) it starts:
> The ROI are defined by paired of masks...

# About the paper
Can you check the above error message after attempting to create the pdf. It seems something is up with the affiliations. @arfon may be able to advise.
