### 2017-12-22 #436 @Kevin-Mattheus-Moerman
@CPernet @smikhael apologies for the delay in processing this review. I've left one lasts check mark open for when you reply to my comments above.

**Happy holidays** :christmas_tree: :sun_with_face:
