### 2017-12-23 #436 @smikhael
@Kevin-Mattheus-Moerman , thank you for you that helpful feedback, and no worries regarding the delay. I'll address all points in the next few days and update you once done. Happy holidays to all!
