### 2017-12-23 #436 @cMadan
Thank you for the detailed review, @Kevin-Mattheus-Moerman!

Along with the suggestions above, I'd suggest maybe having a flag to enable/disable the figure generation.

After you make these last few edits, I'll give the a code a quick try and see if I have any additional usability suggestions. I think I'm well within the target audience (and do plan to use this for a current project), so hopefully that'll be useful. After that we'll be all set!
