### 2017-12-24 db00aa9
Editing minor typos
Fixing typos in lines 59 ('sagittal'), 68 ('sagittal'), and  177 ('handle')
