+++ masks2metrics.m
@@ -56,7 +56,7 @@ fprintf(['Masks2Metrics  Copyright (C) 2017  S. Mikhael and C. Gray. This progra
 %Request direction along which the region of interest was drawn in a continuous manner
 prompt = ['Along which direction was the region of interest drawn?\n Please note that this ' ...
     'needs to have been done in a continuous manner. Press:'...
-    '\n ''1'' for saggital (i.e., along the x-axis)'...
+    '\n ''1'' for sagittal (i.e., along the x-axis)'...
     '\n ''2'' for coronal (i.e., along the y-axis)'...
     '\n ''3'' for axial (i.e., along the z-axis):  '];
 direction = input(prompt);
@@ -65,7 +65,7 @@ direction = input(prompt);
 while (direction~=1&&direction~=2&&direction~=3)
     prompt = ['\n\nPlease enter a valid number (1/2/3) for the direction in which the ROI has'...
         'been continuously drawn:' ...
-    '\n ''1'' for saggital (i.e., along the x-axis)'...
+    '\n ''1'' for sagittal (i.e., along the x-axis)'...
     '\n ''2'' for coronal (i.e., along the y-axis)'...
     '\n ''3'' for axial (i.e., along the z-axis):  '];
     direction = input(prompt);
@@ -174,7 +174,7 @@ total_wm_sa = total_wm_sa * vox_z;
 %% Write all desired variables to files
 
 %Write the filled roi volume to a nifti file
-%convert array to type double as make_nii cannot hand logical data
+%convert array to type double as make_nii cannot handle logical data
 total_filled_roi=double(total_filled_roi);
 filled_name = strcat(int2str(subj),'_',roi,'_',hem,'_filled_step',int2str(step_size));
 filename_filled=make_nii(total_filled_roi,[1 1 2],[0 0 0],2,'binary file');
