### 2017-12-24 e78db09
deleted saving .xls files and added 'draw' warning
-Deleted the writing of the (thickness, volume and surface area) measurements to .xls files (lines 209-214). This is redundant as we are also writing them to .mat files.
-added a warning at the beginning of the masks2metrics (lines 74-84) which appears when called with 'draw=1'. Warning reminds the user that figures will be generated for every segment, and asking them to quit if they want to set draw to 0.
-printing out total thickness parameters, as well as surface area and volume of the ROI
