+++ masks2metrics.m
@@ -42,7 +42,7 @@ function [roi_gm_mean_thickness, roi_gm_vol, roi_wm_sa] = masks2metrics(subj, se
 %    This program is distributed in the hope that it will be useful,
 %    but WITHOUT ANY WARRANTY; without even the implied warranty of
 %    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-%    GNU General Public License for more details.
+%%    GNU General Public License for more details.
 %
 %    You should have received a copy of the GNU General Public License
 %    along with this program.  If not, see <http://www.gnu.org/licenses/>.
@@ -71,8 +71,19 @@ while (direction~=1&&direction~=2&&direction~=3)
     direction = input(prompt);
 end
 
+%if draw=1, print warning reminding user that figures will be generated and displayed. They can
+%quit or press any key to continue.
+if(draw) %draw=1
+    fprintf(['\n\nWARNING: you have set the variable ''draw'' to 1, so Masks2Metrics',...
+          '\n will generate and display 2 figures for each of the %d segments,',...
+        '\n for every slice in which these segments appear. ',...
+        '\n\nPress:',...
+        '\n\t - Ctrl + C to quit and change ''draw'' to 0, or',...
+        '\n\t - any key to continue'], segments);
+    pause
+end
 
-disp(['Preparing ' roi ' for subject ' num2str(subj) '... ']);
+disp(['\nPreparing ' roi ' for subject ' num2str(subj) '... ']);
  
 %Initializing variables for the entire gyrus: thickness, mean Hausdorff 
 %distance, Frechet's distance
@@ -178,6 +189,7 @@ total_wm_sa = total_wm_sa * vox_z;
 total_filled_roi=double(total_filled_roi);
 filled_name = strcat(int2str(subj),'_',roi,'_',hem,'_filled_step',int2str(step_size));
 filename_filled=make_nii(total_filled_roi,[1 1 2],[0 0 0],2,'binary file');
+fprintf('\nWriting binary representing the filled %s volume...',roi)
 filled_name_nii=strcat(filled_name,'.nii');
 save_nii(filename_filled,filled_name_nii);
 
@@ -191,29 +203,14 @@ total_gm_vol=total_gm_area; % sum of GM areas/slice makes up GMV of that roi
 
 mean_total_f=mean(total_f);
 mean_total_mhd=mean(total_mhd);
+fprintf('\nTotal mean Frechet thickness of %s: %2.2f mm',roi,mean_total_f);
+fprintf('\nTotal mean modified Hausdorff distance of %s: %2.2f mm',roi,mean_total_mhd);
 
 max_total_f = max(total_f);
 max_total_mhd = max(total_mhd);
 
-
-%write thickness data to file
-%th_range = 'A1';
-
-file1 = strcat(int2str(subj),'_',roi,'_',hem,'_thickness_WMtoGM_step',int2str(step_size),'.xls');
-file2 = strcat(int2str(subj),'_',roi,'_',hem,'_thickness_GMtoWM_step',int2str(step_size),'.xls');
-file3 = strcat(int2str(subj),'_',roi,'_',hem,'_frechet_step',int2str(step_size),'.xls');
-file4 = strcat(int2str(subj),'_',roi,'_',hem,'_modified_hausdorff_step',int2str(step_size),'.xls');
-file5 = strcat(int2str(subj),'_',roi,'_',hem,'_WM_sa_step',int2str(step_size),'.xls');
[…]
