### 2017-12-24 6e733d1
deleted display of some outputs
deleted display of nonzero trim mean (line 259) and median (line 264) thicknesses
