### 2017-12-24 195d5c6
saving parameters to files and note to user
- saving total ROI thickness, surface area and volume to files
- printing message to user for referring to wiki for more figure information
