### 2017-12-24 42ef1e4
minor bug fix
