+++ masks2metrics.m
@@ -74,10 +74,10 @@ end
 %if draw=1, print warning reminding user that figures will be generated and displayed. They can
 %quit or press any key to continue.
 if(draw) %draw=1
-    fprintf(['\n\nWARNING: you have set the variable ''draw'' to 1, so Masks2Metrics',...
+    fprintf(['\n\nWARNING: you have set the value of ''draw'' to 1, so Masks2Metrics',...
           '\n will generate and display 2 figures for each of the %d segments,',...
         '\n for every slice in which these segments appear. ',...
-	'\n Please refer to the Wiki for more information on the content of the figures.',...
+	'\n Please refer to the Wiki for more information on figure content.',...
         '\n\nPress:',...
         '\n\t - Ctrl + C to quit and change ''draw'' to 0, or',...
         '\n\t - any key to continue'], segments);
@@ -149,7 +149,9 @@ for seg=1:1:segments
             
     end
    
-
+   if draw==0
+     close gcf %to close figure (bug)
+ end
  
     %After running through all the volume for a given gm/wm segment, append
     %this segment's stats to the current list of all segments stats for this gyrus
