### 2017-12-24 7f38d16
Fixed affiliation for CG
-used quotes around 1,2 as it's more than 1 affiliation
