+++ README.md
@@ -17,7 +17,7 @@ Copyright (C) 2017 S. Mikhael and C. Gray.
 
 Masks2Metrics (M2M) is a Matlab based tool that is used to calculate 3 metrics for a given region-of-interest (ROI) in a 3D image: thickness, volume and suface area. While many software packages that automatically compute such metrics exist, it is also necessary to compare the results from automated software with manually-traced ROIs. The current software takes such manually-traced ROIs and computes the metrics automatically.
 
-We (the authors and contributors) use M2M to compute those metrics on images acquired by a Magnetic Resonance Imaging (MRI) machine. The ROI are defined by pairs of masks (in NIfTI format) defining the inner and outer borders of the ROI, and they are drawn continuously along one direction (x-, y- or z-axis). For the special case of brain images, if the ROI describes a gyrus, the paired masks would be the corresponding grey matter (GM) and white matter curves (WM). Paired ROI NIfTI (.nii) masks are expected to be of the form subj_roi_hem_gm/wmsegments.nii. An example of a pair corresponding to subject 1 right superior frontal gyrus (SFG) would be 1_sfg_r_gm1.nii and 1_sfg_r_wm1.nii. A special feature of M2M is that multiple segments can be used rather than a single continuous ROI (see Wiki help). The ROI metrics calculated are grey matter thickness (GMth), grey matter volume (GMvol),and white matter surface area (WMsa).
+We (the authors and contributors) use M2M to compute those metrics on images acquired by a Magnetic Resonance Imaging (MRI) machine. The ROI is defined by pairs of 3-dimensional (3D) binary masks (in NIfTI format) that represent the inner and outer borders of the ROI, and are drawn continuously along one direction (x-, y- or z-axis). For the special case of brain images, if the ROI describes a gyrus, the paired masks would be the corresponding grey matter (GM) and white matter curves (WM). Paired ROI NIfTI (.nii) masks are expected to be of the form subj_roi_hem_gm/wmsegments.nii. An example of a pair corresponding to subject 1 right superior frontal gyrus (SFG) would be 1_sfg_r_gm1.nii and 1_sfg_r_wm1.nii. A special feature of M2M is that multiple segments can be used rather than a single continuous ROI (see Wiki help). The ROI metrics calculated are grey matter thickness (GMth), grey matter volume (GMvol),and white matter surface area (WMsa).
 
 ## License <a name="License"></a>
 
