### 2017-12-25 #436 @smikhael
@Kevin-Mattheus-Moerman  @cMadan   thank you all for your helpful feedback once again. I have now attempted to address all the points you identified as follows:

* flag to enable/disable figures: there was a flag (last variable <code>draw</code> when calling masks2metrics), but I've now made it much clearer using warnings in both the code and the Wiki. Apologies for all those figures jumping at you. They're quite annoying when undesired!

* Changes to the Wiki:
   * correct, the word <code>function</code> was accidentally pasted in. I've now deleted it.
   * added extra line explaining the ROI, as suggested
   * added link to nifti file format
   * added reference to MRIcron as a tool for drawn nifti files and pointing to their drawing tutorial.
   * added table describing all files generated
   * added 'Additional Figure Information' section detailing the contents of each of the figures
   * added a warning to set the variable/flag 'draw' to 0 in case the user doesn't want figures

* Changes to Masks2Metrics:
  * deleted the writing of the (thickness, volume and surface area) measurements to .xls files (previously lines 209-214). This is both confusing and redundant (used in the early days of the code only) as we are also writing these measurements to .mat files.
  * added a warning at the beginning (lines 74-84) which only appears if the value of the flag 'draw' passed to the function equals 1. If draw=0 then the figures are not generated. The warning reminds the user that figures will be generated for every segment, asks them to quit if they don't want that, and to call Masks2Metrics with flag draw=0. The note also points the user to the wiki for more information on figure content.
  * got it to print out total thickness parameters, as well as surface area and volume of the ROI, and to save (.mat format) those that weren't being saved previously.

* Readme:
  * typo now fixed. New text says:  The ROI is defined by pairs of 3-dimensional (3D) binary masks (in NIfTI format) that represent the inner and outer borders of the ROI,...

* pdf generation:
  * Lastly, I've attempted to fix the affiliations problem of the 2nd author (added " " around 1,2 for multiple affiliations) but I still can't generate a pdf. I believe it can't 'cd' to the paper. @arfon  any advice please?

Please let me know if you'd like me to address any additional issues. I look forward to your feedback.

I hope this proves useful to your work @cMadan  and to that of many others! Let me know if you run into any issues. I'm happy to help.

Thank you all once again.
