### 2017-12-25 #436 @cMadan
Just a quick thought re: the PDF compiling, I think the issue is that there are spaces in "paper for JOSS". Can you try renaming the folder to just "paper" or replacing the spaces with underscores?

I will look over the rest of your responses (and try the toolbox) in the next few days.
