### 2017-12-25 f2098ae
uploading JOSS paper with no spaces in folder name
