+++ JOSS_paper/paper.bib
@@ -1,0 +1,175 @@
+
+@article{Bauer_2013,
+author = {Bauer, Stefan and Wiest, Roland and Nolte, Lutz-P and Reyes, Mauricio},
+year = {2013},
+month = {06},
+pages = {R97-R129},
+title = {A survey of MRI-based medical image analysis for brain tumor studies},
+volume = {58},
+booktitle = {Physics in medicine and biology},
+doi = {10.1088/0031-9155/58/13/R97},
+url = {http://iopscience.iop.org/article/10.1088/0031-9155/58/13/R97/meta}
+}
+
+@article{Desikan_2006,
+title = "An automated labeling system for subdividing the human cerebral cortex on MRI scans into gyral based regions of interest",
+journal = "NeuroImage",
+volume = "31",
+number = "3",
+pages = "968 - 980",
+year = "2006",
+issn = "1053-8119",
+doi = "https://doi.org/10.1016/j.neuroimage.2006.01.021",
+url = "http://www.sciencedirect.com/science/article/pii/S1053811906000437",
+author = "Rahul S. Desikan and Florent Ségonne and Bruce Fischl and Brian T. Quinn and Bradford C. Dickerson and Deborah Blacker and Randy L. Buckner and Anders M. Dale and R. Paul Maguire and Bradley T. Hyman and Marilyn S. Albert and Ronald J. Killiany"
+}
+
+@Article{Dickerson_2011,
+  author = {B. C. Dickerson and T. R. Stoub and R. C. Shah and R. A. Sperling and R. J. Killiany and M. S. Albert and B. T. Hyman and D. Blacker and L. deToledo-Morrell},
+  title = {{Alzheimer-signature {MRI} biomarker predicts {AD} dementia in cognitively normal adults}},
+  journal = {Neurology},
+  volume = {76},
+  number = {16},
+  pages = {1395-1402},
+  month = {APR},
+  year = {2011},
+  doi = {10.1212/wnl.0b013e3182166e96},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/21490323?dopt=Abstract}
+}
+
+@article{Fischl_2000,
+author = {Fischl, Bruce and Dale, Anders M.}, 
+title = {Measuring the thickness of the human cerebral cortex from magnetic resonance images},
+volume = {97}, 
+number = {20}, 
+pages = {11050-11055}, 
+year = {2000}, 
+doi = {10.1073/pnas.200033797}, 
+URL = {http://www.pnas.org/content/97/20/11050.abstract}, 
+eprint = {http://www.pnas.org/content/97/20/11050.full.pdf}, 
+journal = {Proceedings of the National Academy of Sciences} 
+}
+
+
+@article{Fischl_2004,
+author = {Fischl, Bruce and van der Kouwe, André and Destrieux, Christophe and Halgren, Eric and Ségonne, Florent and Salat, David H. and Busa, Evelina and Seidman, Larry J. and Goldstein, Jill and Kennedy, David and Caviness, Verne and Makris, Nikos and Rosen, Bruce and Dale, Anders M.},
+title = {Automatically Parcellating the Human Cerebral Cortex},
+journal = {Cerebral Cortex},
+volume = {14},
+number = {1},
+pages = {11-22},
+year = {2004},
+doi = {10.1093/cercor/bhg087},
+URL = { + http://dx.doi.org/10.1093/cercor/bhg087},
+eprint = {/oup/backfile/content_public/journal/cercor/14/1/10.1093/cercor/bhg087/2/bhg087.pdf}
+}
+
+
+@Article{Kochunov2012,
+author="Kochunov, Peter
+and Rogers, William
+and Mangin, Jean-Francois
[…]
