### 2017-12-25 028d284
deleting paper.md
