### 2017-12-25 3b3ea41
deleting paper.bib
