### 2017-12-25 #436 @Kevin-Mattheus-Moerman
It looks like your references heading does not render well. Perhaps a space is needed
