### 2017-12-26 93065b6
fixed 1 reference and added another
-Kochunov2012 --> Kochunov_2012
-added bib for Mills_2016
