+++ JOSS_paper/paper.bib
@@ -64,7 +64,7 @@ eprint = {/oup/backfile/content_public/journal/cercor/14/1/10.1093/cercor/bhg087
 }
 
 
-@Article{Kochunov2012,
+@Article{Kochunov_2012,
 author="Kochunov, Peter
 and Rogers, William
 and Mangin, Jean-Francois
@@ -123,6 +123,19 @@ ISSN={1662-5161}
   urldate = {2017-10-20}
 }
 
+@article{Mills_2016,
+title = "Structural brain development between childhood and adulthood: Convergence across four longitudinal samples",
+journal = "NeuroImage",
+volume = "141",
+number = "Supplement C",
+pages = "273 - 281",
+year = "2016",
+issn = "1053-8119",
+doi = "https://doi.org/10.1016/j.neuroimage.2016.07.044",
+url = "http://www.sciencedirect.com/science/article/pii/S1053811916303512",
+author = "Kathryn L. Mills and Anne-Lise Goddings and Megan M. Herting and Rosa Meuwese and Sarah-Jayne Blakemore and Eveline A. Crone and Ronald E. Dahl and Berna Güroğlu and Armin Raznahan and Elizabeth R. Sowell and Christian K. Tamnes",
+keywords = "Adolescence, Cerebral cortex, MRI, Replication, Sex differences, White matter"
+}
 
 @online{SasiKanth_2011,
   author = {S SasiKanth},
