### 2017-12-26 #436 @smikhael
Thank you @Kevin-Mattheus-Moerman . I've now added the space, and fixed 2 other referencing issues that I discovered (1. needed '_' between Kochunov and 2012, and 2. added the Mills_2016 in the .bib file.). Apologies for the oversight.  Please let me know if you'd like me to regenerate the pdf.
