### 2017-12-26 #436 @Kevin-Mattheus-Moerman
The paper looks good
