### 2017-12-27 5704984
fixed Mikhael_2017 reference
Fixed authors structuring for the reference Mikhael_2017.
