+++ JOSS_paper/paper.bib
@@ -116,7 +116,7 @@ ISSN={1662-5161}
 }
 
 @online{Mikhael_2017,
-  author = {Shadia Mikhael, Calum Gray},
+  author = {Mikhael, Shadia and Gray, Calum},
   title = {Masks2Metrics},
   year = 2017,
   url = {https://github.com/Edinburgh-Imaging/Masks2Metrics},
