### 2017-12-27 #436 @smikhael
Apologies, but I've just regenerated pdf as I uncovered a mistake in one reference (the GitHub reference to our code). The way I had written it made it appear as though the reference only had one author (me) rather than 2. The pdf should now be typo-free. Thank you.
