### 2018-02-12 #436 @smikhael
Excellent, thank you! @cMadan , I archived it just last week on Edinburgh DataShare at https://datashare.is.ed.ac.uk/handle/10283/3018, with the doi: 10.7488/ds/2302. Is that any good or do you prefer another archive on Zenodo/FigShare?
