### 2018-02-14 #436 @smikhael
Thank you all!  I've now added a link to the paper from the README.
